Adding the project
Did the 5 or so mandatory things, added a few quality of life changes. Can enable/disable if a legal move is allowed
Can clear the board so that you can make your own board
Added two play the game for me options, one that does it so you win the other is just to fill the board.
Made a set as default option, I made the way saves work such that we are always working with two files. One that is being worked on then one that we can always go back towards. Ideally, they save and the default would rarely be the same, but users.